requirements.txt
Cython==0.29.24
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
python-dotenv==0.19.1
